Added the main java and database(and connector) code
Includes the code for the first login frame window and the home frame window in UserLogin.java and Home.java respectively. Adds the background images for each tab(or category of words). Also includes initialization and connector code for the database. Added the SQL code for creating the tables for all the words and their particulars.

Create README.md